+++ src/cow-react/modules/tokensList/updaters/TokensListUpdater.ts
@@ -1,7 +1,7 @@
 import { useUpdateAtom } from 'jotai/utils'
 import { tokensByAddressAtom, tokensBySymbolAtom, TokenWithLogo } from '@cow/modules/tokensList/state/tokensListAtom'
 import { useEffect } from 'react'
-import { useTokensListWithDefaults } from 'state/lists/hooks/hooksMod'
+import { useTokensListWithDefaults } from 'state/lists/hooks'
 
 /**
  * This updater protects from redundant recalculations
