--- src/custom/state/lists/hooks/index.ts
@@ -1,11 +1,0 @@
-export * from '@src/state/lists/hooks'
-export {
-  useActiveListUrls,
-  useIsListActive,
-  useAllLists,
-  useGpUnsupportedTokens,
-  useAddGpUnsupportedToken,
-  useIsUnsupportedTokenGp,
-  useRemoveGpUnsupportedToken,
-  useInactiveListUrls,
-} from './hooksMod'
