--- src/custom/state/lists/reducer/index.test.ts
@@ -1,1 +1,0 @@
-export * from './reducerMod.test'
