+++ src/state/lists/reducer.test.ts
@@ -1,10 +1,11 @@
 import { createStore, Store } from 'redux'
+import { DEFAULT_ACTIVE_LIST_URLS_BY_NETWORK, DEFAULT_LIST_OF_LISTS_BY_NETWORK } from 'constants/lists'
+import { updateVersion } from 'state/global/actions'
+import { fetchTokenList, acceptListUpdate, addList, removeList, enableList } from 'state/lists/actions'
+import reducer, { ListsStateByNetwork } from 'state/lists/reducer'
+import { SupportedChainId as ChainId } from '@cowprotocol/cow-sdk'
 
-import { DEFAULT_LIST_OF_LISTS } from 'constants/lists'
-import { DEFAULT_ACTIVE_LIST_URLS } from '../../constants/lists'
-import { updateVersion } from '../global/actions'
-import { acceptListUpdate, addList, enableList, fetchTokenList, removeList } from './actions'
-import reducer, { ListsState } from './reducer'
+const DEFAULT_LIST_OF_LISTS = DEFAULT_LIST_OF_LISTS_BY_NETWORK[ChainId.MAINNET]
 
 const STUB_TOKEN_LIST = {
   name: '',
@@ -26,21 +27,41 @@ const MAJOR_UPDATED_STUB_LIST = {
   version: { ...STUB_TOKEN_LIST.version, major: STUB_TOKEN_LIST.version.major + 1 },
 }
 
+const EMPTY_STORE = {
+  [ChainId.MAINNET]: {
+    byUrl: {},
+    activeListUrls: undefined,
+    gpUnsupportedTokens: {},
+  },
+  [ChainId.GOERLI]: {
+    byUrl: {},
+    activeListUrls: undefined,
+    gpUnsupportedTokens: {},
+  },
+  [ChainId.GNOSIS_CHAIN]: {
+    byUrl: {},
+    activeListUrls: undefined,
+    gpUnsupportedTokens: {},
+  },
+}
+
 describe('list reducer', () => {
-  let store: Store<ListsState>
+  let store: Store<ListsStateByNetwork>
+  const getStoreSlice = (chainId: ChainId) => store.getState()[chainId]
 
   beforeEach(() => {
-    store = createStore(reducer, {
-      byUrl: {},
-      activeListUrls: undefined,
-    })
+    store = createStore(reducer, EMPTY_STORE)
   })
 
-  describe('fetchTokenList', () => {
+  describe('fetchTokenList [MAINNET]', () => {
+    // set what chain we're testing
+    const CHAIN_ID = ChainId.MAINNET
+
     describe('pending', () => {
       it('sets pending', () => {
-        store.dispatch(fetchTokenList.pending({ requestId: 'request-id', url: 'fake-url' }))
-        expect(store.getState()).toEqual({
+        store.dispatch(fetchTokenList.pending({ chainId: CHAIN_ID, requestId: 'request-id', url: 'fake-url' }))
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
+          activeListUrls: undefined,
           byUrl: {
             'fake-url': {
               error: null,
@@ -50,24 +71,29 @@ describe('list reducer', () => {
             },
           },
           selectedListUrl: undefined,
+          gpUnsupportedTokens: {},
         })
       })
 
       it('does not clear current list', () => {
         store = createStore(reducer, {
-          byUrl: {
-            'fake-url': {
-              error: null,
-              current: STUB_TOKEN_LIST,
-              pendingUpdate: null,
-              loadingRequestId: null,
+          ...EMPTY_STORE,
+          [CHAIN_ID]: {
+            byUrl: {
+              'fake-url': {
+                error: null,
+                current: STUB_TOKEN_LIST,
+                pendingUpdate: null,
+                loadingRequestId: null,
+              },
             },
+            activeListUrls: undefined,
+            gpUnsupportedTokens: {},
           },
-          activeListUrls: undefined,
         })
 
         store.dispatch(fetchTokenList.pending({ requestId: 'request-id', url: 'fake-url' }))
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {
             'fake-url': {
               error: null,
@@ -77,6 +103,7 @@ describe('list reducer', () => {
             },
           },
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
     })
@@ -84,9 +111,14 @@ describe('list reducer', () => {
     describe('fulfilled', () => {
       it('saves the list', () => {
         store.dispatch(
-          fetchTokenList.fulfilled({ tokenList: STUB_TOKEN_LIST, requestId: 'request-id', url: 'fake-url' })
+          fetchTokenList.fulfilled({
+            chainId: CHAIN_ID,
+            tokenList: STUB_TOKEN_LIST,
+            requestId: 'request-id',
+            url: 'fake-url',
+          })
         )
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {
             'fake-url': {
               error: null,
@@ -96,6 +128,7 @@ describe('list reducer', () => {
             },
           },
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
 
@@ -106,7 +139,7 @@ describe('list reducer', () => {
         store.dispatch(
           fetchTokenList.fulfilled({ tokenList: STUB_TOKEN_LIST, requestId: 'request-id', url: 'fake-url' })
         )
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {
             'fake-url': {
               error: null,
@@ -116,6 +149,7 @@ describe('list reducer', () => {
             },
           },
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
 
@@ -127,7 +161,7 @@ describe('list reducer', () => {
         store.dispatch(
           fetchTokenList.fulfilled({ tokenList: PATCHED_STUB_LIST, requestId: 'request-id', url: 'fake-url' })
         )
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {
             'fake-url': {
               error: null,
@@ -137,6 +171,7 @@ describe('list reducer', () => {
             },
           },
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
       it('does not save to current if list is newer minor version', () => {
@@ -147,7 +182,7 @@ describe('list reducer', () => {
         store.dispatch(
           fetchTokenList.fulfilled({ tokenList: MINOR_UPDATED_STUB_LIST, requestId: 'request-id', url: 'fake-url' })
         )
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {
             'fake-url': {
               error: null,
@@ -157,6 +192,7 @@ describe('list reducer', () => {
             },
           },
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
       it('does not save to pending if list is newer major version', () => {
@@ -167,7 +203,7 @@ describe('list reducer', () => {
         store.dispatch(
           fetchTokenList.fulfilled({ tokenList: MAJOR_UPDATED_STUB_LIST, requestId: 'request-id', url: 'fake-url' })
         )
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {
             'fake-url': {
               error: null,
@@ -177,6 +213,7 @@ describe('list reducer', () => {
             },
           },
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
     })
@@ -184,26 +221,31 @@ describe('list reducer', () => {
     describe('rejected', () => {
       it('no-op if not loading', () => {
         store.dispatch(fetchTokenList.rejected({ requestId: 'request-id', errorMessage: 'abcd', url: 'fake-url' }))
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {},
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
 
       it('sets the error if loading', () => {
         store = createStore(reducer, {
-          byUrl: {
-            'fake-url': {
-              error: null,
-              current: null,
-              loadingRequestId: 'request-id',
-              pendingUpdate: null,
+          ...EMPTY_STORE,
+          [CHAIN_ID]: {
+            byUrl: {
+              'fake-url': {
+                error: null,
+                current: null,
+                loadingRequestId: 'request-id',
+                pendingUpdate: null,
+              },
             },
+            activeListUrls: undefined,
+            gpUnsupportedTokens: {},
           },
-          activeListUrls: undefined,
         })
         store.dispatch(fetchTokenList.rejected({ requestId: 'request-id', errorMessage: 'abcd', url: 'fake-url' }))
-        expect(store.getState()).toEqual({
+        expect(getStoreSlice(CHAIN_ID)).toEqual({
           byUrl: {
             'fake-url': {
               error: 'abcd',
@@ -213,15 +255,17 @@ describe('list reducer', () => {
             },
           },
           activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         })
       })
     })
   })
 
   describe('addList', () => {
+    const CHAIN_ID = ChainId.MAINNET
     it('adds the list key to byUrl', () => {
-      store.dispatch(addList('list-id'))
-      expect(store.getState()).toEqual({
+      store.dispatch(addList({ url: 'list-id', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {
           'list-id': {
             error: null,
@@ -231,22 +275,27 @@ describe('list reducer', () => {
           },
         },
         activeListUrls: undefined,
+        gpUnsupportedTokens: {},
       })
     })
     it('no op for existing list', () => {
       store = createStore(reducer, {
-        byUrl: {
-          'fake-url': {
-            error: null,
-            current: STUB_TOKEN_LIST,
-            loadingRequestId: null,
-            pendingUpdate: null,
+        ...EMPTY_STORE,
+        [CHAIN_ID]: {
+          byUrl: {
+            'fake-url': {
+              error: null,
+              current: STUB_TOKEN_LIST,
+              loadingRequestId: null,
+              pendingUpdate: null,
+            },
           },
+          activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         },
-        activeListUrls: undefined,
       })
-      store.dispatch(addList('fake-url'))
-      expect(store.getState()).toEqual({
+      store.dispatch(addList({ url: 'fake-url', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {
           'fake-url': {
             error: null,
@@ -256,25 +305,31 @@ describe('list reducer', () => {
           },
         },
         activeListUrls: undefined,
+        gpUnsupportedTokens: {},
       })
     })
   })
 
   describe('acceptListUpdate', () => {
+    const CHAIN_ID = ChainId.MAINNET
     it('swaps pending update into current', () => {
       store = createStore(reducer, {
-        byUrl: {
-          'fake-url': {
-            error: null,
-            current: STUB_TOKEN_LIST,
-            loadingRequestId: null,
-            pendingUpdate: PATCHED_STUB_LIST,
+        ...EMPTY_STORE,
+        [CHAIN_ID]: {
+          byUrl: {
+            'fake-url': {
+              error: null,
+              current: STUB_TOKEN_LIST,
+              loadingRequestId: null,
+              pendingUpdate: PATCHED_STUB_LIST,
+            },
           },
+          activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         },
-        activeListUrls: undefined,
       })
-      store.dispatch(acceptListUpdate('fake-url'))
-      expect(store.getState()).toEqual({
+      store.dispatch(acceptListUpdate({ url: 'fake-url', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {
           'fake-url': {
             error: null,
@@ -284,64 +339,81 @@ describe('list reducer', () => {
           },
         },
         activeListUrls: undefined,
+        gpUnsupportedTokens: {},
       })
     })
   })
 
   describe('removeList', () => {
+    const CHAIN_ID = ChainId.MAINNET
     it('deletes the list key', () => {
       store = createStore(reducer, {
-        byUrl: {
-          'fake-url': {
-            error: null,
-            current: STUB_TOKEN_LIST,
-            loadingRequestId: null,
-            pendingUpdate: PATCHED_STUB_LIST,
+        ...EMPTY_STORE,
+        [CHAIN_ID]: {
+          byUrl: {
+            'fake-url': {
+              error: null,
+              current: STUB_TOKEN_LIST,
+              loadingRequestId: null,
+              pendingUpdate: PATCHED_STUB_LIST,
+            },
           },
+          activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         },
-        activeListUrls: undefined,
       })
-      store.dispatch(removeList('fake-url'))
-      expect(store.getState()).toEqual({
+      store.dispatch(removeList({ url: 'fake-url', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {},
         activeListUrls: undefined,
+        gpUnsupportedTokens: {},
       })
     })
     it('Removes from active lists if active list is removed', () => {
       store = createStore(reducer, {
-        byUrl: {
-          'fake-url': {
-            error: null,
-            current: STUB_TOKEN_LIST,
-            loadingRequestId: null,
-            pendingUpdate: PATCHED_STUB_LIST,
+        ...EMPTY_STORE,
+        [CHAIN_ID]: {
+          byUrl: {
+            'fake-url': {
+              error: null,
+              current: STUB_TOKEN_LIST,
+              loadingRequestId: null,
+              pendingUpdate: PATCHED_STUB_LIST,
+            },
           },
+          activeListUrls: ['fake-url'],
+          gpUnsupportedTokens: {},
         },
-        activeListUrls: ['fake-url'],
       })
-      store.dispatch(removeList('fake-url'))
-      expect(store.getState()).toEqual({
+      store.dispatch(removeList({ url: 'fake-url', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {},
         activeListUrls: [],
+        gpUnsupportedTokens: {},
       })
     })
   })
 
   describe('enableList', () => {
+    const CHAIN_ID = ChainId.MAINNET
     it('enables a list url', () => {
       store = createStore(reducer, {
-        byUrl: {
-          'fake-url': {
-            error: null,
-            current: STUB_TOKEN_LIST,
-            loadingRequestId: null,
-            pendingUpdate: PATCHED_STUB_LIST,
+        ...EMPTY_STORE,
+        [CHAIN_ID]: {
+          byUrl: {
+            'fake-url': {
+              error: null,
+              current: STUB_TOKEN_LIST,
+              loadingRequestId: null,
+              pendingUpdate: PATCHED_STUB_LIST,
+            },
           },
+          activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         },
-        activeListUrls: undefined,
       })
-      store.dispatch(enableList('fake-url'))
-      expect(store.getState()).toEqual({
+      store.dispatch(enableList({ url: 'fake-url', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {
           'fake-url': {
             error: null,
@@ -351,22 +423,27 @@ describe('list reducer', () => {
           },
         },
         activeListUrls: ['fake-url'],
+        gpUnsupportedTokens: {},
       })
     })
     it('adds to url keys if not present already on enable', () => {
       store = createStore(reducer, {
-        byUrl: {
-          'fake-url': {
-            error: null,
-            current: STUB_TOKEN_LIST,
-            loadingRequestId: null,
-            pendingUpdate: PATCHED_STUB_LIST,
+        ...EMPTY_STORE,
+        [CHAIN_ID]: {
+          byUrl: {
+            'fake-url': {
+              error: null,
+              current: STUB_TOKEN_LIST,
+              loadingRequestId: null,
+              pendingUpdate: PATCHED_STUB_LIST,
+            },
           },
+          activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         },
-        activeListUrls: undefined,
       })
-      store.dispatch(enableList('fake-url-invalid'))
-      expect(store.getState()).toEqual({
+      store.dispatch(enableList({ url: 'fake-url-invalid', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {
           'fake-url': {
             error: null,
@@ -382,22 +459,27 @@ describe('list reducer', () => {
           },
         },
         activeListUrls: ['fake-url-invalid'],
+        gpUnsupportedTokens: {},
       })
     })
     it('enable works if list already added', () => {
       store = createStore(reducer, {
-        byUrl: {
-          'fake-url': {
-            error: null,
-            current: null,
-            loadingRequestId: null,
-            pendingUpdate: null,
+        ...EMPTY_STORE,
+        [CHAIN_ID]: {
+          byUrl: {
+            'fake-url': {
+              error: null,
+              current: null,
+              loadingRequestId: null,
+              pendingUpdate: null,
+            },
           },
+          activeListUrls: undefined,
+          gpUnsupportedTokens: {},
         },
-        activeListUrls: undefined,
       })
-      store.dispatch(enableList('fake-url'))
-      expect(store.getState()).toEqual({
+      store.dispatch(enableList({ url: 'fake-url', chainId: CHAIN_ID }))
+      expect(getStoreSlice(CHAIN_ID)).toEqual({
         byUrl: {
           'fake-url': {
             error: null,
@@ -407,45 +489,53 @@ describe('list reducer', () => {
           },
         },
         activeListUrls: ['fake-url'],
+        gpUnsupportedTokens: {},
       })
     })
   })
 
   describe('updateVersion', () => {
+    const CHAIN_ID = ChainId.MAINNET
     describe('never initialized', () => {
       beforeEach(() => {
         store = createStore(reducer, {
-          byUrl: {
-            'https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json': {
-              error: null,
-              current: STUB_TOKEN_LIST,
-              loadingRequestId: null,
-              pendingUpdate: null,
+          ...EMPTY_STORE,
+          [CHAIN_ID]: {
+            byUrl: {
+              'https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json': {
+                error: null,
+                current: STUB_TOKEN_LIST,
+                loadingRequestId: null,
+                pendingUpdate: null,
+              },
+              'https://unpkg.com/@uniswap/default-token-list@latest': {
+                error: null,
+                current: STUB_TOKEN_LIST,
+                loadingRequestId: null,
+                pendingUpdate: null,
+              },
             },
-            'https://unpkg.com/@uniswap/default-token-list@latest': {
-              error: null,
-              current: STUB_TOKEN_LIST,
-              loadingRequestId: null,
-              pendingUpdate: null,
-            },
+            activeListUrls: undefined,
+            gpUnsupportedTokens: {},
           },
-          activeListUrls: undefined,
         })
-        store.dispatch(updateVersion())
+        store.dispatch(updateVersion({ chainId: CHAIN_ID }))
       })
 
       it('clears the current lists', () => {
         expect(
-          store.getState().byUrl['https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json']
+          getStoreSlice(CHAIN_ID).byUrl[
+            'https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json'
+          ]
         ).toBeUndefined()
-        expect(store.getState().byUrl['https://unpkg.com/@uniswap/default-token-list@latest']).toBeUndefined()
+        expect(getStoreSlice(CHAIN_ID).byUrl['https://unpkg.com/@uniswap/default-token-list@latest']).toBeUndefined()
       })
 
       it('puts in all the new lists', () => {
-        expect(Object.keys(store.getState().byUrl)).toEqual(DEFAULT_LIST_OF_LISTS)
+        expect(Object.keys(getStoreSlice(CHAIN_ID).byUrl)).toEqual(DEFAULT_LIST_OF_LISTS)
       })
       it('all lists are empty', () => {
-        const s = store.getState()
+        const s = getStoreSlice(CHAIN_ID)
         Object.keys(s.byUrl).forEach((url) => {
           expect(s.byUrl[url]).toEqual({
             error: null,
@@ -456,38 +546,44 @@ describe('list reducer', () => {
         })
       })
       it('sets initialized lists', () => {
-        expect(store.getState().lastInitializedDefaultListOfLists).toEqual(DEFAULT_LIST_OF_LISTS)
+        expect(getStoreSlice(CHAIN_ID).lastInitializedDefaultListOfLists).toEqual(DEFAULT_LIST_OF_LISTS)
       })
       it('sets selected list', () => {
-        expect(store.getState().activeListUrls).toEqual(DEFAULT_ACTIVE_LIST_URLS)
+        expect(getStoreSlice(CHAIN_ID).activeListUrls).toEqual(DEFAULT_ACTIVE_LIST_URLS_BY_NETWORK[CHAIN_ID])
       })
     })
     describe('initialized with a different set of lists', () => {
       beforeEach(() => {
         store = createStore(reducer, {
-          byUrl: {
-            'https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json': {
-              error: null,
-              current: STUB_TOKEN_LIST,
-              loadingRequestId: null,
-              pendingUpdate: null,
+          ...EMPTY_STORE,
+          [CHAIN_ID]: {
+            byUrl: {
+              'https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json': {
+                error: null,
+                current: STUB_TOKEN_LIST,
+                loadingRequestId: null,
+                pendingUpdate: null,
+              },
+              'https://unpkg.com/@uniswap/default-token-list@latest': {
+                error: null,
+                current: STUB_TOKEN_LIST,
+                loadingRequestId: null,
+                pendingUpdate: null,
+              },
             },
-            'https://unpkg.com/@uniswap/default-token-list@latest': {
-              error: null,
-              current: STUB_TOKEN_LIST,
-              loadingRequestId: null,
-              pendingUpdate: null,
-            },
+            activeListUrls: undefined,
+            gpUnsupportedTokens: {},
+            lastInitializedDefaultListOfLists: ['https://unpkg.com/@uniswap/default-token-list@latest'],
           },
-          activeListUrls: undefined,
-          lastInitializedDefaultListOfLists: ['https://unpkg.com/@uniswap/default-token-list@latest'],
         })
-        store.dispatch(updateVersion())
+        store.dispatch(updateVersion({ chainId: CHAIN_ID }))
       })
 
       it('does not remove lists not in last initialized list of lists', () => {
         expect(
-          store.getState().byUrl['https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json']
+          getStoreSlice(CHAIN_ID).byUrl[
+            'https://unpkg.com/@uniswap/default-token-list@latest/uniswap-default.tokenlist.json'
+          ]
         ).toEqual({
           error: null,
           current: STUB_TOKEN_LIST,
@@ -496,11 +592,11 @@ describe('list reducer', () => {
         })
       })
       it('removes lists in the last initialized list of lists', () => {
-        expect(store.getState().byUrl['https://unpkg.com/@uniswap/default-token-list@latest']).toBeUndefined()
+        expect(getStoreSlice(CHAIN_ID).byUrl['https://unpkg.com/@uniswap/default-token-list@latest']).toBeUndefined()
       })
 
       it('each of those initialized lists is empty', () => {
-        const byUrl = store.getState().byUrl
+        const byUrl = getStoreSlice(CHAIN_ID).byUrl
         // note we don't expect the uniswap default list to be prepopulated
         // this is ok.
         Object.keys(byUrl).forEach((url) => {
@@ -516,10 +612,10 @@ describe('list reducer', () => {
       })
 
       it('sets initialized lists', () => {
-        expect(store.getState().lastInitializedDefaultListOfLists).toEqual(DEFAULT_LIST_OF_LISTS)
+        expect(getStoreSlice(CHAIN_ID).lastInitializedDefaultListOfLists).toEqual(DEFAULT_LIST_OF_LISTS)
       })
       it('sets default list to selected list', () => {
-        expect(store.getState().activeListUrls).toEqual(DEFAULT_ACTIVE_LIST_URLS)
+        expect(getStoreSlice(CHAIN_ID).activeListUrls).toEqual(DEFAULT_ACTIVE_LIST_URLS_BY_NETWORK[CHAIN_ID])
       })
     })
   })
